Move a bunch more classes/resources out of the main common lib
