+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/message/MessagePointer.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.message
+package uk.ac.wellcome.messaging.message
 
 import java.net.URI
