+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/sns/NotificationMessage.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.sns
+package uk.ac.wellcome.messaging.sns
 
 case class NotificationMessage(
   MessageId: String,
