+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/sqs/SQSReader.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.sqs
+package uk.ac.wellcome.messaging.sqs
 
 import com.amazonaws.services.sqs.AmazonSQS
 import com.amazonaws.services.sqs.model.{
