+++ sbt_common/messaging/test/scala/uk/ac/wellcome/messaging/sqs/SQSWorkerToDynamoTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.sqs
+package uk.ac.wellcome.messaging.sqs
 
 import akka.actor.ActorSystem
 import com.amazonaws.services.cloudwatch.AmazonCloudWatch
