+++ sbt_common/messaging/test/scala/uk/ac/wellcome/messaging/test/fixtures/SQS.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.test.fixtures
+package uk.ac.wellcome.messaging.test.fixtures
 
 import com.amazonaws.auth._
 import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration
